테스트 안하고 올려서 죄송ㅠ_ ㅠ.. 테스트 완 디비 수정

+++ MyApplication/app/src/main/java/com/example/cds/eattle_prototype_2/StoryListAdapter.java
@@ -4,14 +4,12 @@ import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Environment;
-import android.text.Layout;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import java.util.ArrayList;
 
@@ -59,7 +57,7 @@ public class StoryListAdapter extends BaseAdapter{
             storyImage.setImageURI(Uri.parse(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM)+"/"+ "thumbnail" +"/"+items.get(position).getImgID()+".jpg"));
 
             //특정 아이템에 해당하는 폴더 아이디를 가져온다
-            final int folderID = items.get(position).getFolderID();
+            final long folderID = items.get(position).getFolderID();
             //리스트 클릭 리스너
             convertView.setOnClickListener(new View.OnClickListener() {
                 @Override
